% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burden_Effect_Size_meta.R
\name{Burden_Effect_Size_meta}
\alias{Burden_Effect_Size_meta}
\title{Effect size and standard error estimates of meta-analysis of burden test for a given variant-set}
\usage{
Burden_Effect_Size_meta(obj_MetaSTAAR_merge, rv_num_cutoff = 2)
}
\arguments{
\item{obj_MetaSTAAR_merge}{an object from merging the summary statistics
and covariance files from each participating study, which is the output from
\code{\link{MetaSTAAR_merge}}.}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}
}
\value{
a list with the following members:

\code{num_variant}: the number of variants with combined minor allele frequency > 0 and less than
\code{rare_maf_cutoff} in the given variant-set that are used for performing the
variant-set test using MetaSTAAR.

\code{cMAC}: the combined cumulative minor allele count of variants with
combined minor allele frequency > 0 and less than \code{rare_maf_cutoff} in the given variant-set.

\code{Burden_Score_Stat}: the score statistic of Burden-MS(1,1) for the given variant-set.

\code{Burden_SE_Score}: the standard error of \code{Burden_Score_Stat} for the given variant-set.

\code{Burden_pvalue}: the Burden-MS(1,1) p-value for the given variant-set.

\code{Burden_Est}: the effect size estimate of Burden-MS(1,1) for the given variant-set.

\code{Burden_SE_Est}: the standard error estimate of \code{Burden_Est} for the given variant-set.
}
\description{
The \code{Burden_Effect_Size_meta} function takes in the object from the merged summary statistics
and covariance files of each participating study and functional annotation data
to calculate the effect size and standard error estimates of meta-analysis of burden test for a given variant-set.
}
\references{
Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics 52}(9), 969-983.
(\href{https://www.nature.com/articles/s41588-020-0676-4}{pub})

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Human Genetics 104}(3), 410-421.
(\href{https://www.sciencedirect.com/science/article/pii/S0002929719300023}{pub})
}
