% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiv_Score_Test_Region_meta_cond.R
\name{Indiv_Score_Test_Region_meta_cond}
\alias{Indiv_Score_Test_Region_meta_cond}
\title{Meta-analysis of conditional score test for individual variants in a given variant-set}
\usage{
Indiv_Score_Test_Region_meta_cond(obj_MetaSTAAR_merge_cond, rv_num_cutoff = 2)
}
\arguments{
\item{obj_MetaSTAAR_merge_cond}{an object from merging the conditional summary statistics
and covariance files from each participating study, adjusting for a given list of variants,
which is the output from \code{\link{MetaSTAAR_merge_cond}}.}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}
}
\value{
a data frame with p rows corresponding to the p genetic variants in the given variant-set
and three columns: \code{Score_cond} (the conditional score test statistic),
\code{SE_cond} (the standard error associated with the conditional score test statistic),
and \code{pvalue_cond} (the conditional score test p-value). If a variant in
the given variant-set has standard error equal to 0, the p-value will be set as 1.
}
\description{
The \code{Indiv_Score_Test_Region_meta_cond} function takes in the object from the
merged conditional summary statistics and covariance files of each participating study to analyze
the conditional associations between a quantitative/dichotomous phenotype and
all individual variants in a given variant-set by using the meta-analysis of score test,
adjusting for a given list of variants.
}
