% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSTAAR_cond.R
\name{MetaSTAAR_cond}
\alias{MetaSTAAR_cond}
\title{Meta-analysis of STAAR (MetaSTAAR) procedure for conditional analysis using omnibus test}
\usage{
MetaSTAAR_cond(
  obj_MetaSTAAR_merge_cond,
  annotation_phred = NULL,
  rv_num_cutoff = 2
)
}
\arguments{
\item{obj_MetaSTAAR_merge_cond}{an object from merging the conditional summary statistics
and covariance files from each individual study, adjusting for a given list of variants,
which is the output from \code{\link{MetaSTAAR_merge_cond}}.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p),
where p is the number of genetic variants in the variant-set.
Continuous scores should be given in PHRED score scale, where the PHRED score
of j-th variant is defined to be -10*log10(rank(-score_j)/total) across the genome. (Binary)
categorical scores should be taking values 0 or 1, where 1 is functional and 0 is
non-functional. If not provided, MetaSTAAR will perform the
SKAT-MS(1,25), SKAT-MS(1,1), Burden-MS(1,25), Burden-MS(1,1), ACAT-V-MS(1,25), ACAT-V-MS(1,1)
and ACAT-O-MS tests (default = NULL).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}
}
\value{
a list with the following members:

\code{num_variant}: the number of variants with combined minor allele frequency > 0 and less than
\code{rare_maf_cutoff} in the given variant-set that are used for performing the
variant-set test using MetaSTAAR.

\code{cMAC}: the combined cumulative minor allele count of variants with
combined minor allele frequency > 0 and less than \code{rare_maf_cutoff} in the given variant-set.

\code{results_MetaSTAAR_O_cond}: the conditional MetaSTAAR-O p-value that aggregated conditional
SKAT-MS(1,25), SKAT-MS(1,1), Burden-MS(1,25), Burden-MS(1,1), ACAT-V-MS(1,25), and ACAT-V-MS(1,1) together
with conditional p-values of each test weighted by each annotation using Cauchy method.

\code{results_ACAT_O_MS_cond}: the conditional ACAT-O-MS p-value that aggregated conditional
SKAT-MS(1,25), SKAT-MS(1,1), Burden-MS(1,25), Burden-MS(1,1), ACAT-V-MS(1,25), and ACAT-V-MS(1,1) using Cauchy method.

\code{results_MetaSTAAR_S_1_25_cond}: a vector of conditional MetaSTAAR-S(1,25) p-values,
including conditional SKAT-MS(1,25) p-value weighted by MAF, the conditional SKAT-MS(1,25)
p-values weighted by each annotation, and a conditional MetaSTAAR-S(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_MetaSTAAR_S_1_1_cond}: a vector of conditional MetaSTAAR-S(1,1) p-values,
including conditional SKAT-MS(1,1) p-value weighted by MAF, the conditional SKAT-MS(1,1)
p-values weighted by each annotation, and a conditional MetaSTAAR-S(1,1)
p-value by aggregating these p-values using Cauchy method.

\code{results_MetaSTAAR_B_1_25_cond}: a vector of conditional MetaSTAAR-B(1,25) p-values,
including conditional Burden-MS(1,25) p-value weighted by MAF, the conditional Burden-MS(1,25)
p-values weighted by each annotation, and a conditional MetaSTAAR-B(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_MetaSTAAR_B_1_1_cond}: a vector of conditional MetaSTAAR-B(1,1) p-values,
including conditional Burden-MS(1,1) p-value weighted by MAF, the conditional Burden-MS(1,1)
p-values weighted by each annotation, and a conditional MetaSTAAR-B(1,1)
p-value by aggregating these p-values using Cauchy method.

\code{results_MetaSTAAR_A_1_25_cond}: a vector of conditional MetaSTAAR-A(1,25) p-values,
including conditional ACAT-V-MS(1,25) p-value weighted by MAF, the conditional ACAT-V-MS(1,25)
p-values weighted by each annotation, and a conditional MetaSTAAR-A(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_MetaSTAAR_A_1_1_cond}: a vector of conditional MetaSTAAR-A(1,1) p-values,
including conditional ACAT-V-MS(1,1) p-value weighted by MAF, the conditional ACAT-V-MS(1,1)
p-values weighted by each annotation, and a conditional MetaSTAAR-A(1,1)
p-value by aggregating these p-values using Cauchy method.
}
\description{
The \code{MetaSTAAR_cond} function takes in the object from the merged conditional summary statistics
and covariance files of each individual study and functional annotation data
to analyze the conditional association between a quantitative/dichotomous phenotype and
a variant-set by using the meta-analysis of STAAR (MetaSTAAR) procedure,
adjusting for a given list of variants.
For each variant-set, the MetaSTAAR-O p-value is a p-value from an omnibus test that aggregated
conditional SKAT-MS(1,25), SKAT-MS(1,1), Burden-MS(1,25), Burden-MS(1,1), ACAT-V-MS(1,25), and ACAT-V-MS(1,1)
together with conditional p-values of each test weighted by each annotation using Cauchy method.
}
\references{
Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics 52}(9), 969-983.
(\href{https://www.nature.com/articles/s41588-020-0676-4}{pub})

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Human Genetics 104}(3), 410-421.
(\href{https://www.sciencedirect.com/science/article/pii/S0002929719300023}{pub})
}
