% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSTAAR_individual_analysis.R
\name{MetaSTAAR_individual_analysis}
\alias{MetaSTAAR_individual_analysis}
\title{Meta-analysis of individual variants using \code{MetaSTAAR}}
\usage{
MetaSTAAR_individual_analysis(
  chr,
  start.loc,
  end.loc,
  study.names,
  sample.sizes,
  sumstat.dir,
  common_mac_cutoff,
  trait,
  segment.size = 5e+05,
  check_qc_label = FALSE
)
}
\arguments{
\item{chr}{a numeric value indicating the chromosome of the genetic region of interest.}

\item{start.loc}{a numeric value indicating the starting location (position) of the
genetic region of interest.}

\item{end.loc}{a numeric value indicating the ending location (position) of the
genetic region of interest.}

\item{study.names}{a character vector containing the name of each individual
study contributing to the meta-analysis.}

\item{sample.sizes}{a numeric vector with the length of \code{study.names}
indicating the sample size of each study.}

\item{sumstat.dir}{a character vector containing the directories of the study-specific summary statistics file folders.}

\item{common_mac_cutoff}{the cutoff of minimum combined minor allele count (inclusive) in
defining "common" variants.}

\item{trait}{a character value indicating the underlying trait of interest for
the meta-analysis.}

\item{segment.size}{a numeric value indicating the length of each segment of which
the summary statistics and sparse weighted covariance files are stored.
Note that the input value should be aligned with the input values of
\code{\link{MetaSTAAR_worker_cov}} (default = 5e+05).}

\item{check_qc_label}{a logical value indicating whether variants need to be dropped according to \code{qc_label}
specified in \code{\link{MetaSTAAR_worker_sumstat}} and \code{\link{MetaSTAAR_worker_cov}}.
If \code{check_qc_label} is FALSE, it is assumed that no variant will be dropped (Default = FALSE).}
}
\value{
a data frame with p rows corresponding to the p genetic variants in the merged variant list
and the following columns: chromosome (chr), position (pos), reference allele (ref), alternative allele (alt),
combined alternative allele count (alt_AC), combined minor allele count (MAC), combined minor allele frequency (MAF),
combined sample size (N), the score test p-value (p), the log score test p-value (logp), the score test statistic (Score),
the standard error associated with the score test statistic (Score_SE), the estimated effect size of the minor allele (Est),
the standard error associated with the estimated effect size (Est_se).
If a variant in the merged variant list has standard error equal to 0, the p-value will be set as 1.
}
\description{
The \code{MetaSTAAR_individual_analysis} function takes in the summary statistics file
from each individual study (the output from \code{\link{MetaSTAAR_worker_sumstat}})
to analyze the associations between a quantitative/dichotomous phenotype and
all individual variants in the merged variant list by using the meta-analysis
}
