% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSTAAR_worker_sumstat.R
\name{MetaSTAAR_worker_sumstat}
\alias{MetaSTAAR_worker_sumstat}
\title{Generating summary statistics file using MetaSTAAR_worker}
\usage{
MetaSTAAR_worker_sumstat(genotype, obj_nullmodel, variant_info)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm}} function for unrelated samples or
\code{\link{fit_null_glmmkin}} function for related samples in the \code{\link{STAAR}} package.}

\item{variant_info}{a data frame or matrix of variant information (unique identifier)
with p rows (listed in the same order as the columns of \code{genotype}) and should contain
the following 4 columns: chromosome (chr), position (pos), reference allele (ref), and alternative allele (alt).}
}
\value{
\code{sumstat}: the data frame of all variants in the variant-set (the summary statistics file),
including the following information: chromosome (chr), position (pos), reference allele (ref),
alternative allele (alt), alternative allele count (alt_AC), minor allele count (MAC),
minor allele frequency (MAF), study sample size (N), score statistic (U), variance (V), and
the (low-rank decomposed) dense component of the covariance file.
}
\description{
The \code{MetaSTAAR_worker_sumstat} function takes in genotype, the object
from fitting the null model, and variant information (unique identifier)
to generate the summary statistics file for the given variant-set.
}
